+++ src/test/java/io/github/tobyjamesclements/parsley/ParsleyProcessorSourceLayerTest.java
@@ -345,6 +345,7 @@ class ParsleyProcessorSourceLayerTest {
             context.addStateStore(bufferStore);
             context.addStateStore(candidateIndexStore);
             context.addStateStore(forwardedIndexStore);
+        context.addStateStore(new ParsleyCommittedCompleteness("frontier-commit-hook"));
             processor.init(context);
         }